Fixes #27 to accept the new event update format
The event scanner doesn't send the tim_author_id when processing the
events. The scanner only sends the service_id, service_author_id and
service_event_id. The event updater uses the service_id and
service_author_id to get the tim_author_id. This is a partial change the
ideal change would completely remove the requirement on needed a
tim_author_id.