Fixes #22. collectors can listen to multiple queue
The collector drivers now listen to multiple queues. This should make it
much easier to deploy as all notification message can be handle by one
process. To execute the collector against all the queue simply:

python event-collector/driver.py

If you want to run the collector only against one queue then run it as:

python event-collector/driver.py --service facebook

To make this change possible the low-level message_queue module was
changed. send_messages now only take two parameters (the client and the
list of message). send_messages is smart enough to route each message to
the correct queue.

join now only take two parameters (the client and a list of handlers).
Each element in the list of handler is a map with two keys ('queue' and
'handler'). The 'queue' key stores the name of the queue associated with
this handler and the 'handler' key stores the closure that should get
called.