Update #35. Aggresively decrease the priority
Changed the scanner so that we aggressively decrease the priority of any
update event that we send. This will allow us to more quickly lower the
priority of events and mitigate the risk of overrunning the updater and
processor.

The processor was also changed so that it never lowers the priority of
an event. It instead only sets the priority to the highest priority
(zero) whenever it detects a change.