Update #23. Refactor AppBase for the daemon.
Clean up all the code around AppBase in preparation for the
implemetnation of daemon support:

1. Remove all the code dealing with positional arguments in AppBase. It
is now the individual app's responsibility to validate the positional
args.

2. Minimized the used of intance variable. Most objections are now
passed in the method called. This removed the dependencies that
applications need to know the implementation of AppBase

3. Split def main into two methods. The application is responsible for
implementing app_main. AppBase implements main and contains
functionality common to all applications.

4. Fix all the application to use the new abstraction